+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fixed the Drill dye recipes not functioning.
 * Fixed the name of the first drill speed augment.
 * Fixed the farmer not funneling seeds back to the seed slots.
+* Added crops block-tag to Canola/Flax/Rice
 
 # 1.3.11+mc1.21.1
 * Fixed Farmer not playing well with non-vanilla farmland.